+++ tests/unit.test.ts
@@ -86,7 +86,7 @@ describe("Arithmetics", () => {
 		expect(fn({ x: 1, y: 2, z: 3 }, [ 1, 2, 3 ])).toBe("(2, 4, 6)");
 	});
 	it("Performs the scalar multiplication of the vector", () => {
-		const fn = (input1: Input, input2: number) => vector(input1).mul(input2).toString();
+		const fn = (input1: Input, value: number) => vector(input1).scale(value).toString();
 
 		expect(fn([ 1 ], 3)).toBe("(3, 0, 0)");
 		expect(fn([ 1, 2 ], 3)).toBe("(3, 6, 0)");
@@ -106,16 +106,6 @@ describe("Arithmetics", () => {
 		expect(fn({ x: 1, y: 2 }, 0.5)).toBe("(0.5, 1, 0)");
 		expect(fn({ x: 1, y: 2 }, 0.5)).toBe("(0.5, 1, 0)");
 		expect(fn({ x: 1, y: 2, z: 3 }, 0.5)).toBe("(0.5, 1, 1.5)");
-	});
-	it("Performs the vector multiplication", () => {
-		const fn = (input1: Input, input2: Input) => vector(input1).mul(input2).toString();
-
-		expect(fn([ 2 ], [ 3 ])).toBe("(6, 0, 0)");
-		expect(fn([ 1, 2 ], [ 3, 4 ])).toBe("(3, 8, 0)");
-		expect(fn([ 1, 2, 3 ], [ 4, 5, 6 ])).toBe("(4, 10, 18)");
-		expect(fn({ x: 2 }, { x: 3 })).toBe("(6, 0, 0)");
-		expect(fn({ x: 1, y: 2 }, { x: 3, y: 4 })).toBe("(3, 8, 0)");
-		expect(fn({ x: 1, y: 2, z: 3 }, { x: 4, y: 5, z: 6 })).toBe("(4, 10, 18)");
 	});
 	it("Inverts the vector", () => {
 		const fn = (input: Input) => vector(input).inverted.toString();