+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Unreleased (2022-11-24)
 
 - [feat]: `.add()` method for vector addition;
-- [feat]: `.mul()` method for vector multiplication;
+- [feat]: `.scale()` method for scalar vector multiplication;
 - [feat]: `inverted` property for inverting the vector;
 
 ## 0.1.0 (2022-11-23)